Create 8x_Combo_PON_OLT_AM
Adtran 4187518F32

The 10G Combo PON OLT module supports a flexible optics approach that allows for each of the eight ports to independently utilize multiple types of transceivers. This allows service providers the flexibility to determine the right technology (GPON or XGS-PON) for residential FTTH while also providing capacity for high-bandwidth business and mobile backhaul services. The module is compatible with the Total Access 5000 Series.