Eaton PSU, EBM and PDUs (#1035)
* Create 9PX5KIRTN.yaml

Just created this one.
In realety the eth0 should be a module which can be fitted in but this is just right for the most applications i think.

* Update 9PX5KIRTN.yaml

Made changes nameing power outlets. Also set "power_port" for the outlets. Added Battery connector and detection interfaces.

* Create 9PXEBM180.yaml

* Create ESWB05.yaml

* Create ESWB22.yaml

* Changed in-/out-put to hardwired

* Create CBL2OUT32.yaml

* Update 9PX5KIRTN.yaml

Fixes to formatting

* Update 9PX5KIRTN.yaml

Removed un-needed definitions

* Update 9PXEBM180.yaml

Removed un-needed definitions

Co-authored-by: Daniel W. Anner <daniel.anner@danstechsupport.com>