Add GS108PEv3 (#1230)
* Add GS180PEv3

I did something wrong with the previous pull request, so I created a new one for this. Appended with additional info.
No airflow as it's a passive switch. No u_height as it's a desktop 8 port poe switch which is not likely to be mounted in a rack.

* Update GS108PEv3.yaml

Fixed slug
Added passive airflow per comment from author
Fixed u_height
Fixed comment
Removed un-needed definitions

---------

Co-authored-by: Daniel W. Anner <daniel.anner@danstechsupport.com>