Update device-types/Inter-Tech/3U-30255.yaml
Please use the same importer in the webfrontend and in your pull requests .. or pin a message at the top of the forum with instructions "for idiots".

BUT the best thing would be - use a file format for adults! --> JSON

The childish stuff with YAML cost me several hours.
This was my first and last contribution to the Netbox community.
It's a shame that such big hurdles are being put in place here.